refactor(threads): useThreadActionsResumeThread 拆 legacyFallback 与 recoveryProbe
- 2293→1394 行：legacyFallback 670（8 个 per-engine 分支）+ recoveryProbe 428
- 文本锁测试先改写为跨文件合并断言（同前缀兄弟文件），拆分后 3/3 绿
- dsh host.down 双 catch 早退语义锁面未动；deps 数组逐字
- baseline 重生成 + 新文件登记豁免
验证：tsc 本文件族零 error；resume-guard+policy 9 passed；claude-history 失败与基线一致（mock 既有）
（子代理 EncouragingMammal 执行，主代理复核验证）
