+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-08-30T20:27:22.271Z",
+  "generatedAt": "2026-08-30T20:28:51.074Z",
   "scope": "fail",
   "policyVersion": "2026-08-30.policy-v5",
   "entries": [
